+++ theme/woostarter-child/functions.php
@@ -1,0 +1,5 @@
+<?php
+
+foreach ( glob( get_stylesheet_directory() . '/inc/*.php' ) as $file ) {
+	require_once $file;
+}